requirements.txt
alembic==1.8.1
amqp==5.0.6
anyio==3.6.2
apache-airflow==2.5.0
apache-airflow-providers-amazon==6.2.0
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.0.0
asn1crypto==1.5.1
attrs==22.1.0
Babel==2.11.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.5
boto3==1.26.41
botocore==1.29.41
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
clipboard==0.0.4
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.32
croniter==1.3.8
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flower==1.0.0
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
humanize==3.12.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
inflection==0.5.1
iso8601==1.0.0
isodate==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==4.17.3
kombu==5.2.1
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
mypy-boto3-appflow==1.26.32
mypy-boto3-rds==1.26.39
mypy-boto3-redshift-data==1.26.30
numpy==1.24.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.5.2
pathspec==0.9.0
pendulum==2.1.2
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
ply==3.11
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psutil==5.9.4
psycopg2-binary==2.9.5
py4j==0.10.9.5
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.2
pyspark==3.3.1
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.909
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
s3transfer==0.6.0
scramp==1.4.4
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.12
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.1
text-unidecode==1.3
tornado==6.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.13
vine==5.0.0
watchtower==2.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.11.0
